Day06
Overriding the method from parent class(super class) to child class(sub class).
for overriding the method data type and args must be same.
To override a method it should be a non static method. for static method we can not override it.
For private class we can not override a method.